+++ test/test_exception_handler.rb
@@ -1,4 +1,4 @@
-require 'helper'
+require_relative 'helper'
 require 'sidekiq/exception_handler'
 require 'stringio'
 require 'logger'