+++ Changes.md
@@ -1,10 +1,12 @@
 HEAD
 -----------
-
+- Add random integer to process identity [#2113, michaeldiscala]
 - Log Sidekiq Pro's Batch ID if available [#2076]
 - Refactor Processor Redis usage to avoid redis/redis-rb#490 [#]
 - Move /dashboard/stats to /stats.  Add /stats/queues. [moserke, #2099]
-
+- Add better usage text for `sidekiqctl`.
+- `Sidekiq::Logging.with_context` is now a stack so you can set your
+  own job context for logging purposes [grosser, #2110]
 
 3.3.0
 -----------